+++ perf/tests/test_perf_cli.py
@@ -186,8 +186,9 @@ class TestPerfCLI(unittest.TestCase):
         expected = ("""
             Number of samples: 250 (50 runs x 5 samples; 1 warmup)
             Standard deviation / median: 1%
-            Shortest raw sample: 264 ms (10 loops)
-            Longest raw sample: 273 ms (10 loops)
+            Loop iterations per sample: 10
+            Raw sample minimum: 264 ms
+            Raw sample maximum: 273 ms
 
             Minimum: 26.4 ms (-2%)
             Median +- std dev: 26.9 ms +- 0.2 ms